+++ atcg_lib/platform/opengl/src/Renderer/Renderer.cpp
@@ -701,13 +701,26 @@ void RendererSystem::draw(const atcg::ref_ptr<Graph>& mesh,
 {
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
 
+    draw(mesh, impl->standard_material, camera, model, color, shader, draw_mode);
+}
+
+void RendererSystem::draw(const atcg::ref_ptr<Graph>& mesh,
+                          const Material& material,
+                          const atcg::ref_ptr<Camera>& camera,
+                          const glm::mat4& model,
+                          const glm::vec3& color,
+                          const atcg::ref_ptr<Shader>& shader,
+                          DrawMode draw_mode)
+{
+    ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
+
     mesh->unmapAllPointers();
     switch(draw_mode)
     {
         case ATCG_DRAW_MODE_TRIANGLE:
         {
             shader->setInt("entityID", -1);
-            impl->setMaterial(impl->standard_material, shader);
+            impl->setMaterial(material, shader);
             impl->drawVAO(mesh->getVerticesArray(),
                           camera,
                           color,
@@ -720,16 +733,14 @@ void RendererSystem::draw(const atcg::ref_ptr<Graph>& mesh,
         case ATCG_DRAW_MODE_POINTS:
         {
             shader->setInt("entityID", -1);
-            impl->setMaterial(impl->standard_material, shader);
-            SystemRegistry::instance()
-                ->getSystem<RendererSystem>()
-                ->impl->drawVAO(mesh->getVerticesArray(), camera, color, shader, model, GL_POINTS, mesh->n_vertices());
+            impl->setMaterial(material, shader);
+            impl->drawVAO(mesh->getVerticesArray(), camera, color, shader, model, GL_POINTS, mesh->n_vertices());
         }
         break;
         case ATCG_DRAW_MODE_POINTS_SPHERE:
         {
             shader->setInt("entityID", -1);
-            impl->setMaterial(impl->standard_material, shader);
+            impl->setMaterial(material, shader);
             impl->drawPointCloudSpheres(mesh->getVerticesArray()->peekVertexBuffer(),
                                         camera,
                                         model,
@@ -742,7 +753,7 @@ void RendererSystem::draw(const atcg::ref_ptr<Graph>& mesh,
         {
             auto edge_shader = impl->shader_manager->getShader("edge");
             edge_shader->setInt("entityID", -1);
-            impl->setMaterial(impl->standard_material, edge_shader);
+            impl->setMaterial(material, edge_shader);
             atcg::ref_ptr<VertexBuffer> points = mesh->getVerticesBuffer();
             points->bindStorage(0);
             impl->drawVAO(mesh->getEdgesArray(), camera, color, edge_shader, model, GL_POINTS, mesh->n_edges(), 1);
@@ -752,20 +763,18 @@ void RendererSystem::draw(const atcg::ref_ptr<Graph>& mesh,
         {
             auto edge_shader = impl->shader_manager->getShader("cylinder_edge");
             edge_shader->setInt("entityID", -1);
-            impl->setMaterial(impl->standard_material, edge_shader);
+            impl->setMaterial(material, edge_shader);
             impl->drawGrid(mesh->getVerticesBuffer(), mesh->getEdgesBuffer(), edge_shader, camera, model, color);
         }
         break;
         case ATCG_DRAW_MODE_INSTANCED:
         {
             shader->setInt("entityID", -1);
-            impl->setMaterial(impl->standard_material, shader);
+            impl->setMaterial(material, shader);
             atcg::ref_ptr<VertexArray> vao_mesh      = mesh->getVerticesArray();
             atcg::ref_ptr<VertexBuffer> instance_vbo = vao_mesh->peekVertexBuffer();
             uint32_t n_instances                     = instance_vbo->size() / instance_vbo->getLayout().getStride();
-            SystemRegistry::instance()
-                ->getSystem<RendererSystem>()
-                ->impl->drawVAO(vao_mesh, camera, color, shader, model, GL_TRIANGLES, mesh->n_vertices(), n_instances);
+            impl->drawVAO(vao_mesh, camera, color, shader, model, GL_TRIANGLES, mesh->n_vertices(), n_instances);
         }
         break;
     }
@@ -947,7 +956,19 @@ void RendererSystem::draw(const atcg::ref_ptr<Scene>& scene, const atcg::ref_ptr
 {
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
 
-    // TODO: Just raw opengl rendering code here
+    drawSkybox(camera);
+
+    const auto& view = scene->getAllEntitiesWith<atcg::TransformComponent>();
+
+    for(auto e: view)
+    {
+        Entity entity(e, scene.get());
+        RendererSystem::draw(entity, camera);
+    }
+}
+
+void RendererSystem::drawSkybox(const atcg::ref_ptr<Camera>& camera)
+{
     if(impl->has_skybox)
     {
         glDepthMask(GL_FALSE);
@@ -964,14 +985,6 @@ void RendererSystem::draw(const atcg::ref_ptr<Scene>& scene, const atcg::ref_ptr
         glDepthMask(GL_TRUE);
         glDepthFunc(GL_LESS);
         toggleCulling(culling);
-    }
-
-    const auto& view = scene->getAllEntitiesWith<atcg::TransformComponent>();
-
-    for(auto e: view)
-    {
-        Entity entity(e, scene.get());
-        RendererSystem::draw(entity, camera);
     }
 }